moved everything document related to the doc search client (to the service, out of the controller)
