+++ 02 - Web UI Template/CognitiveSearch.UI/Search/DocumentSearchClient.cs
@@ -5,20 +5,18 @@ using Microsoft.ApplicationInsights;
 using Microsoft.Azure.Search;
 using Microsoft.Azure.Search.Models;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Spatial;
-using Newtonsoft.Json.Linq;
 using System;
 using System.Collections.Generic;
-using System.Configuration;
-using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
-using System.Web;
+using Microsoft.WindowsAzure.Storage.Blob;
+using Microsoft.WindowsAzure.Storage.Auth;
 
 namespace CognitiveSearch.UI
 {
     public class DocumentSearchClient
     {
+        private IConfiguration _configuration { get; set; }
         private SearchServiceClient _searchClient;
         private ISearchIndexClient _indexClient;
         private string searchServiceName { get; set; }
@@ -37,10 +35,21 @@ namespace CognitiveSearch.UI
 
         public static string errorMessage;
 
+        bool _isPathBase64Encoded { get; set; }
+
+        // data source information. Currently supporting 3 data sources indexed by different indexers
+        private static string[] containerAddresses = null;
+        private static string[] tokens = null;
+
+        // this should match the default value used in appsettings.json.  
+        private static string defaultContainerUriValue = "https://{storage-account-name}.blob.core.windows.net/{container-name}";
+
+
         public DocumentSearchClient(IConfiguration configuration)
         {
             try
             {
+                _configuration = configuration;
                 searchServiceName = configuration.GetSection("SearchServiceName")?.Value;
                 apiKey = configuration.GetSection("SearchApiKey")?.Value;
                 IndexName = configuration.GetSection("SearchIndexName")?.Value;
@@ -54,6 +63,8 @@ namespace CognitiveSearch.UI
 
                 Schema = new SearchSchema().AddFields(_searchClient.Indexes.Get(IndexName).Fields);
                 Model = new SearchModel(Schema);
+
+                _isPathBase64Encoded = (configuration.GetSection("IsPathBase64Encoded")?.Value == "True");
 
             }
             catch (Exception e)
@@ -246,6 +257,62 @@ namespace CognitiveSearch.UI
             return null;
         }
 
+        public DocumentResult GetDocuments(string q, SearchFacet[] searchFacets, int currentPage)
+        {
+            var tokens = GetContainerSasUris();
+
+            var selectFilter = Model.SelectFilter;
+
+            if (!string.IsNullOrEmpty(q))
+            {
+                q = q.Replace("?", "");
+            }
+
+            var response = Search(q, searchFacets, selectFilter, currentPage);
+            var searchId = GetSearchId().ToString();
+            var facetResults = new List<object>();
+            var tagsResults = new List<object>();
+
+            if (response != null && response.Facets != null)
+            {
+                // Return only the selected facets from the Search Model
+                foreach (var facetResult in response.Facets.Where(f => Model.Facets.Where(x => x.Name == f.Key).Any()))
+                {
+                    var cleanValues = GetCleanFacetValues(facetResult);
+
+                    facetResults.Add(new
+                    {
+                        key = facetResult.Key,
+                        value = cleanValues
+                    });
+                }
+
+                foreach (var tagResult in response.Facets.Where(t => Model.Tags.Where(x => x.Name == t.Key).Any()))
+                {
+                    var cleanValues = GetCleanFacetValues(tagResult);
+
+                    tagsResults.Add(new
+                    {
+                        key = tagResult.Key,
+                        value = cleanValues
+                    });
+                }
+            }
+
+            var result = new DocumentResult
+            {
+                Results = (response == null ? null : response.Results),
+                Facets = facetResults,
+                Tags = tagsResults,
+                Count = (response == null ? 0 : Convert.ToInt32(response.Count)),
+                SearchId = searchId,
+                IdField = idField,
+                Token = tokens[0],
+                IsPathBase64Encoded = _isPathBase64Encoded
+            };
+            return result;
+        }
+
         /// <summary>
         /// Initiates a run of the search indexer.
         /// </summary>
@@ -255,6 +322,160 @@ namespace CognitiveSearch.UI
             if (indexStatus.LastResult.Status != IndexerExecutionStatus.InProgress)
             {
                 _searchClient.Indexers.Run(IndexerName);
+            }
+        }
+
+        private string GetToken(string decodedPath)
+        {
+            // Initialize tokens and containers if not already initialized
+            GetContainerSasUris();
+
+            // Determine which token to use.
+            string tokenToUse;
+            if (decodedPath.ToLower().Contains(containerAddresses[1])) { tokenToUse = tokens[1]; }
+            else if (decodedPath.ToLower().Contains(containerAddresses[2])) { tokenToUse = tokens[2]; }
+            else { tokenToUse = tokens[0]; }
+
+            return tokenToUse;
+        }
+
+        /// <summary>
+        /// This will return up to 3 tokens for the storage accounts
+        /// </summary>
+        /// <returns></returns>
+        private string[] GetContainerSasUris()
+        {
+            // We need to refresh the tokens every time or they will become invalid.
+            tokens = new string[3];
+            containerAddresses = new string[3];
+
+            string accountName = _configuration.GetSection("StorageAccountName")?.Value;
+            string accountKey = _configuration.GetSection("StorageAccountKey")?.Value;
+
+            SharedAccessBlobPolicy adHocPolicy = new SharedAccessBlobPolicy()
+            {
+                SharedAccessExpiryTime = DateTime.UtcNow.AddHours(24),
+                Permissions = SharedAccessBlobPermissions.Read
+            };
+
+            containerAddresses[0] = _configuration.GetSection("StorageContainerAddress")?.Value.ToLower();
+            CloudBlobContainer container = new CloudBlobContainer(new Uri(containerAddresses[0]), new StorageCredentials(accountName, accountKey));
+            tokens[0] = container.GetSharedAccessSignature(adHocPolicy, null);
+
+            // Get token for second indexer data source
+            containerAddresses[1] = _configuration.GetSection("StorageContainerAddress2")?.Value.ToLower();
+            if (!String.Equals(containerAddresses[1], defaultContainerUriValue))
+            {
+                CloudBlobContainer container2 = new CloudBlobContainer(new Uri(containerAddresses[1]), new StorageCredentials(accountName, accountKey));
+                tokens[1] = container2.GetSharedAccessSignature(adHocPolicy, null);
+            }
+
+            // Get token for third indexer data source
+            containerAddresses[2] = _configuration.GetSection("StorageContainerAddress3")?.Value.ToLower();
+            if (!String.Equals(containerAddresses[2], defaultContainerUriValue))
+            {
+                CloudBlobContainer container3 = new CloudBlobContainer(new Uri(containerAddresses[2]), new StorageCredentials(accountName, accountKey));
+                tokens[2] = container3.GetSharedAccessSignature(adHocPolicy, null);
+            }
+
+            return tokens;
+        }
+
+        public DocumentResult GetDocumentById(string id)
+        {
+            var decodedPath = id;
+
+            var response = LookUp(id);
+
+            if (_isPathBase64Encoded)
+            {
+                decodedPath = Base64Decode(id);
+            }
+
+            string tokenToUse = GetToken(decodedPath);
+
+            var result = new DocumentResult
+            {
+                Result = response,
+                Token = tokenToUse,
+                DecodedPath = decodedPath,
+                IdField = idField,
+                IsPathBase64Encoded = _isPathBase64Encoded
+            };
+            return result;
+        }
+
+        private static string Base64Decode(string input)
+        {
+            if (input == null) throw new ArgumentNullException("input");
+            int inputLength = input.Length;
+            if (inputLength < 1) return null;
+
+            // Get padding chars
+            int numPadChars = (int)input[inputLength - 1] - (int)'0';
+            if (numPadChars < 0 || numPadChars > 10)
+            {
+                return null;
+            }
+
+            // replace '-' and '_'
+            char[] base64Chars = new char[inputLength - 1 + numPadChars];
+            for (int iter = 0; iter < inputLength - 1; iter++)
+            {
+                char c = input[iter];
+
+                switch (c)
+                {
+                    case '-':
+                        base64Chars[iter] = '+';
+                        break;
+
+                    case '_':
+                        base64Chars[iter] = '/';
+                        break;
+
+                    default:
+                        base64Chars[iter] = c;
+                        break;
+                }
+            }
+
+            // Add padding chars
+            for (int iter = inputLength - 1; iter < base64Chars.Length; iter++)
+            {
+                base64Chars[iter] = '=';
+            }
+
+            var charArray = Convert.FromBase64CharArray(base64Chars, 0, base64Chars.Length);
+            return System.Text.Encoding.Default.GetString(charArray);
+        }
+
+        /// <summary>
+        /// In some situations you may want to restrict the facets that are displayed in the
+        /// UI. This allows you to add some heuristics to remove facets that you may consider unnecessary.
+        /// </summary>
+        /// <param name="facetResult"></param>
+        /// <returns></returns>
+        private static IList<FacetResult> GetCleanFacetValues(KeyValuePair<string, IList<FacetResult>> facetResult)
+        {
+            IList<FacetResult> cleanValues = new List<FacetResult>();
+
+            if (facetResult.Key == "people")
+            {
+                // only add names that are long enough 
+                foreach (var element in facetResult.Value)
+                {
+                    if (element.Value.ToString().Length >= 4)
+                    {
+                        cleanValues.Add(element);
+                    }
+                }
+
+                return cleanValues;
+            }
+            else
+            {
+                return facetResult.Value;
             }
         }
     }
